directory structure setting moved to general settings
